FavoriteScreen which re-uses the ImageViewer Component
+-------+-------------------------------------------------------------------------+
| Type  | Task Description                                                        |
+-------+-------------------------------------------------------------------------+
| Item  | Create a Favorite button in the HomeScreen to access the                |
|       | FavoriteScreen                                                          |
+-------+-------------------------------------------------------------------------+
| Item  | Create a FavoriteScreen which re-uses the ImageViewer Component from    |
|       | GalleryScreen                                                           |
+-------+-------------------------------------------------------------------------+
